# proto/searches.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.36.11
// 	protoc        v3.21.12
// source: proto/searches.proto

package proto

import (
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
	unsafe "unsafe"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type GetTopNRequest struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	N             int64                  `protobuf:"varint,1,opt,name=n,proto3" json:"n,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *GetTopNRequest) Reset() {
	*x = GetTopNRequest{}
	mi := &file_proto_searches_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *GetTopNRequest) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*GetTopNRequest) ProtoMessage() {}

func (x *GetTopNRequest) ProtoReflect() protoreflect.Message {
	mi := &file_proto_searches_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use GetTopNRequest.ProtoReflect.Descriptor instead.
func (*GetTopNRequest) Descriptor() ([]byte, []int) {
	return file_proto_searches_proto_rawDescGZIP(), []int{0}
}

func (x *GetTopNRequest) GetN() int64 {
	if x != nil {
		return x.N
	}
	return 0
}

type GetTopNResponse struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	TopSearches   []*TopSearch           `protobuf:"bytes,1,rep,name=top_searches,json=topSearches,proto3" json:"top_searches,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *GetTopNResponse) Reset() {
	*x = GetTopNResponse{}
	mi := &file_proto_searches_proto_msgTypes[1]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *GetTopNResponse) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*GetTopNResponse) ProtoMessage() {}

func (x *GetTopNResponse) ProtoReflect() protoreflect.Message {
	mi := &file_proto_searches_proto_msgTypes[1]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use GetTopNResponse.ProtoReflect.Descriptor instead.
func (*GetTopNResponse) Descriptor() ([]byte, []int) {
	return file_proto_searches_proto_rawDescGZIP(), []int{1}
}

func (x *GetTopNResponse) GetTopSearches() []*TopSearch {
	if x != nil {
		return x.TopSearches
	}
	return nil
}

type TopSearch struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Search        string                 `protobuf:"bytes,1,opt,name=search,proto3" json:"search,omitempty"`
	Count         int64                  `protobuf:"varint,2,opt,name=count,proto3" json:"count,omitempty"`
	LastSearched  string                 `protobuf:"bytes,3,opt,name=last_searched,json=lastSearched,proto3" json:"last_searched,omitempty"` // ISO 8601
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *TopSearch) Reset() {
	*x = TopSearch{}
	mi := &file_proto_searches_proto_msgTypes[2]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *TopSearch) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*TopSearch) ProtoMessage() {}

func (x *TopSearch) ProtoReflect() protoreflect.Message {
	mi := &file_proto_searches_proto_msgTypes[2]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use TopSearch.ProtoReflect.Descriptor instead.
func (*TopSearch) Descriptor() ([]byte, []int) {
	return file_proto_searches_proto_rawDescGZIP(), []int{2}
}

func (x *TopSearch) GetSearch() string {
	if x != nil {
		return x.Search
	}
	return ""
}

func (x *TopSearch) GetCount() int64 {
	if x != nil {
		return x.Count
	}
	return 0
}

func (x *TopSearch) GetLastSearched() string {
	if x != nil {
		return x.LastSearched
	}
	return ""
}

var File_proto_searches_proto protoreflect.FileDescriptor

const file_proto_searches_proto_rawDesc = "" +
	"\n" +
	"\x14proto/searches.proto\x12\bsearches\"\x1e\n" +
	"\x0eGetTopNRequest\x12\f\n" +
	"\x01n\x18\x01 \x01(\x03R\x01n\"I\n" +
	"\x0fGetTopNResponse\x126\n" +
	"\ftop_searches\x18\x01 \x03(\v2\x13.searches.TopSearchR\vtopSearches\"^\n" +
	"\tTopSearch\x12\x16\n" +
	"\x06search\x18\x01 \x01(\tR\x06search\x12\x14\n" +
	"\x05count\x18\x02 \x01(\x03R\x05count\x12#\n" +
	"\rlast_searched\x18\x03 \x01(\tR\flastSearched2J\n" +
	"\bSearches\x12>\n" +
	"\aGetTopN\x12\x18.searches.GetTopNRequest\x1a\x19.searches.GetTopNResponseB8Z6github.com/middelmatigheid/wb-top-searches/proto;protob\x06proto3"

var (
	file_proto_searches_proto_rawDescOnce sync.Once
	file_proto_searches_proto_rawDescData []byte
)

func file_proto_searches_proto_rawDescGZIP() []byte {
	file_proto_searches_proto_rawDescOnce.Do(func() {
		file_proto_searches_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_proto_searches_proto_rawDesc), len(file_proto_searches_proto_rawDesc)))
	})
	return file_proto_searches_proto_rawDescData
}

var file_proto_searches_proto_msgTypes = make([]protoimpl.MessageInfo, 3)
var file_proto_searches_proto_goTypes = []any{
	(*GetTopNRequest)(nil),  // 0: searches.GetTopNRequest
	(*GetTopNResponse)(nil), // 1: searches.GetTopNResponse
	(*TopSearch)(nil),       // 2: searches.TopSearch
}
var file_proto_searches_proto_depIdxs = []int32{
	2, // 0: searches.GetTopNResponse.top_searches:type_name -> searches.TopSearch
	0, // 1: searches.Searches.GetTopN:input_type -> searches.GetTopNRequest
	1, // 2: searches.Searches.GetTopN:output_type -> searches.GetTopNResponse
	2, // [2:3] is the sub-list for method output_type
	1, // [1:2] is the sub-list for method input_type
	1, // [1:1] is the sub-list for extension type_name
	1, // [1:1] is the sub-list for extension extendee
	0, // [0:1] is the sub-list for field type_name
}

func init() { file_proto_searches_proto_init() }
func file_proto_searches_proto_init() {
	if File_proto_searches_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: unsafe.Slice(unsafe.StringData(file_proto_searches_proto_rawDesc), len(file_proto_searches_proto_rawDesc)),
			NumEnums:      0,
			NumMessages:   3,
			NumExtensions: 0,
			NumServices:   1,
		},
		GoTypes:           file_proto_searches_proto_goTypes,
		DependencyIndexes: file_proto_searches_proto_depIdxs,
		MessageInfos:      file_proto_searches_proto_msgTypes,
	}.Build()
	File_proto_searches_proto = out.File
	file_proto_searches_proto_goTypes = nil
	file_proto_searches_proto_depIdxs = nil
}